⚠️Huge update, refactored species geodata for multiple countries in a single species
Moved all geo logic to backend for a better approach to location storage and display. Geo-data is now stored in the speces table, and each of its positions are shown as markers.

Position generation is made by using the same dispersion and border-enforcement algorythm for each country the species may be associated with.

This approach ensures that species in multiple countries are handled correctly.