⚠️ Major update, species real-time udpate
When the user clicks on a species, if a set period of time has passed (14 days for now), the backend will update the data on said species using the IUCN and GBIF API endpoints.

In the process, media, description, red list category and common_name will be validated. If any of them exist, the fetching is skipped, except for red list category, which will always be updated.